Rename class select -> select_expr
When `sys/select.h` is included, we get into an ambiguous name situation between the syscall, this class, and the `expr select(expr, expr, expr);` call that is hard to resolve without adding additional `using` calls in multiple places. SInce the macOS version of GTest likes to include that header, this makes for awkward mac-specific compile errors. Renaming this class seems to be the simplest approach to making the code compile more happily.