Clean up alloc-on-stack code for raw_buffer
We repeated ~same logic three times for this pattern, and it's error prone. We can't wrap it cleanly because alloca() won't let us, but wrapping it in a macro and a placement-new at least moves the relevant initialization logic out of evaluate and into buffer itself.

Also: drive-by change to alloca() usage for copy_dims because IMHO it makes for cleaner code.