Partial fix for #8
Replace the existing func::make() calls with a variadic one, with a somewhat thinner and more generic wrapper. As a bonus, the template arguments need not be specified if the callsite is a "real" function... but if it's a std::function or lambda, they still do (I haven't figured out if I can get the argument type inference working for these cases). Not sure if this is a finished product yet but PTAL.