Update performance app to do 2 copies to work around #22 (#47)
This changes the performance app to do two copies via an intermediate
buffer. This works around #22, and also produces data that is a bit more
interesting (but also a bit more complicated).

Reproducing this data requires disabling the `alias_buffers`
optimization, which causes the intermediate buffer to be aliased to the
output, and in turn causes memcpy to behave strangely.

The test when using `memmove` and with `alias_buffers` is also
interesting, but the story gets a lot more complicated still. Basically,
we have the cost of two dispatches to do one operation (the other
memmove is a no-op).