Enable and test `buffer<void>` more (#58)
This mostly already worked, but in order to cast to void, asserts needed
to be relaxed. I realized that these asserts aren't really needed
(unlike either https://github.com/dsharlet/array or Halide), because our
strides are in bytes and not elements, so casts still point at the right
elements.