Speed up build (maybe) (#64)
I noticed that it took over a minute to compile simplify.cc with asan on
GH actions. This PR tries to speed up the build in a few ways.
Unfortunately, none of them make that big of a difference, but they are
good changes anyays.

I tried further splitting up simplify_exprs.cc, but it barely made a
difference. It seems the problem isn't necessarily the volume of
code/expressions, but a particular pathological one I guess?