Padded copies must always have crops (fixes #21) (#85)
This PR makes it so copies that are padded always have a crop that
applies to the output, which defines where the padding will be applied,
and more importantly, limits the bounds required of the source buffer we
are (potentially) adding padding to.

This exposed some bugs in the design of the existing API, the trivial
copy tests become less interesting with this change. What these two
tests did (translate by an input parameter, and rely on `copy` to pad as
needed) relies on the fact that the inputs to the copy were inputs to
the pipeline, and the API that supported that is hard to use if the copy
inputs *aren't* pipeline inputs. So, this PR removes that API entirely.