Add for_each_contiguous_slice_multi(), which is a variadic version of for_each_contiguous_slice().
I added this as a separate API because during dev work I didn't want to involve conflicts, but the API is backwards-compatible and the performance is basically identical when you pass just a single buffer (according to the benchmarks I added).

I think that the fusing and folding handling is basically correct; there might be additional opportunities for optimization (especially for folding) but I think this is ready for review.