Clean up buffer types (fixes #81) (#92)
I realized that we could avoid tricky lifetime issues for
`raw_buffer::make` helpers by allocating the buffer and data all in one
`malloc`ed buffer. Now the result can simply be freed to clean up.

This PR also changes `raw_buffer::allocate` to return a pointer to be
freed manually, instead of storing it in the class. `buffer<T, N>` now
keeps a pointer to be freed, and automatically manages lifetime.