Add `callable_attrs::allow_in_place` and use that for aliasing (#124)
Basically this is #11. I'm not that happy with how this appears in the
`func` API, I wish the attributes were attached to the callbacks
themselves, so the callable would only need to be annotated with its
attributes once, instead of N times at each use.

But, this is mostly a step in that direction, and it allows cleaning up
the alias_buffers optimization quite a bit.