Fix more bugs (#140)
I realized that the way the simplifier handles symbol references isn't
safe, because we see symbol references from nodes that later get
simplified or removed entirely. Fixing this uncovered more bugs. Fixes
in this PR:

- Use `depends_on` instead of trying to avoid recursive visitors in
simplify. This has bad algorithmic complexity, but it's correct. We can
fix the algorithmic complexity later if/when it's a problem.
- Slices need to update *all* references to the buffer metadata, not
just the bounds of the buffer itself.
- Lets need to be simplified in reverse order to get interdependent lets
correct.
- `depends_on` didn't handle shadowing correctly.