Make `pipeline` "editable" by clients (#157)
I have some cases where I'd like to modify the generated pipeline in
some simple ways, e.g. adding extra checks. This is doable by making a
new pipeline with edits (if we expose `constants` which is currently
inaccessible), but I'm not sure the immutable style really makes sense
for this.