Update the initial IR statement construction and improve the logic to compute the innermost locations (#158)
This PR makes initial IR construction to explicitly use the precomputed
topological order + uses the order to compute innermost locations of the
functions without explicit compute_at.

The high-level structure of the code:
1) compute topological order for the functions in the pipeline.
2) using the order from 1) build loop nest tree and compute undefined
compute_at locations.
3) construct the initial IR statement using 1) and 2).

This addresses the issues described in #146 and some other bugs. For the
test from #146, the final IR is nice and clean:

```
intm2 = allocate(heap, 2, {
  {[min(buffer_min(intm4, 0), buffer_min(intm3, 0)), max(buffer_max(intm4, 0), buffer_max(intm3, 0))], 2, <>},
  {[min(buffer_min(intm4, 1), buffer_min(intm3, 1)), max(buffer_max(intm4, 1), buffer_max(intm3, 1))], (((max(buffer_max(intm4, 0), buffer_max(intm3, 0)) - min(buffer_min(intm4, 0), buffer_min(intm3, 0))) * 2) + 2), <>}
}) {
 call(<fn>, {in1}, {intm2})
 call(<fn>, {intm2}, {intm3})
 call(<fn>, {intm2}, {intm4})
}
```

All existing tests pass with  ```bazel test //...```.

Unrelated change: I've added a name field to the func to use it for
easier debugging. I can move it to another PR or remove it altogether if
people think it's not needed.