Reduce dependencies between expr and stmt (#179)
This isn't exactly what I set out to do, but I think it's a nice
improvement in that it reduces some dependencies. This enables us to
avoid the dumb pattern of calling `std::abort()` for every `stmt` in
some visitors for example.

If you want to see changes in this PR in more detail, looking at the
individual commits will be more useful than looking at the entire diff
at once.