Change `elem_size` from `size_t` to `expr` (#206)
The thing that motivated this is now abandoned, but I think this is a
good change anyways. It should not affect performance significantly, it
is more flexible this way, and it eliminates a use of size_t in the API.
`make_buffer` already necessarily uses `expr elem_size` too, so this
makes things more consistent.