Sanitize user expressions of buffer metadata (#219)
This PR avoids a common and tempting footgun: using buffer metadata
expressions in the definition of bounds, crops, concatenate bounds, etc.

The problem with this is that these expressions get placed into the
generated pipeline inside crops, which changes the value of those buffer
metadata expressions. This is almost certainly not what the user
intended.

This PR fixes this by adding a "sanitization" pass that replaces buffer
metadata expressions in user provided expressions with variables that
are then later defined to have the same value as the original buffer
metadata expression, but with the value they have immediately upon
entering the pipeline.

This is effectively equivalent to what we need the user to do anyways:
pass the buffer metadata as a separate scalar parameter to the pipeline,
and use that scalar in their expression definitions. The problem is that
this is very tedious and hard to automate outside of slinky.

This PR includes some refactoring/cleanup related to this, look at
abe58751fd1c4e12745fba18f38dde2a495c01af in isolation to see the
meaningful change.