Clean up `symbol_id` and `var` (#220)
A few things have bothered me for a while:
- Lack of strong typing for `symbol_id`, it allowed doing arithmetic and
implicit conversions to ints (and then maybe `expr` as an integer and
not a variable).
- `var` was a hacky wrapper to fix the first thing.

This PR *maybe* cleans this up by replacing `symbol_id` with `var`. This
means we can now safely define `expr(var)`, and we can get rid of a lot
of boilerplate.