Tighten bounds of some subs with patterns (#237)
There are some expressions that are very hard to simplify that also fail
to produce tight bounds due to limitations of interval arithmetic. This
PR works around this by adding some special cases to such expressions in
`bounds_of`.

This finds the fold factors in the 4 level pipeline. However, doing so
causes asserts regarding cropping and folding to fail (See #40 and #59
for more information). "Fortunately", we don't actually do this folding
on main, because we can only get this folding if we hack `where_true` to
work in this case.