Make simplifier rule canonical checks compile time asserts (#250)
It was incredibly irritating to be working on simplifier rules, and need
to run gdb in order to figure out which rule had a problem with
canonical ordering. This PR makes the check a static_assert, which makes
it easy to see which rule failed at compile time, instead of the stack
at runtime.