Buffer helper improvements (#252)
- `for_each_contiguous_slice` is now type safe (fixes #133)
- Adds `for_each_element`

I think that we should consider removing `for_each_slice` entirely. It's
very hard to make it type safe. Existing uses of it can use
`for_each_element`, and just explicitly slice the leading dimensions as
desired first. The implicit slicing of buffers in `for_each_slice` can
be confusing, because there is some accommodation for missing dimensions
at both ends of the dimension list. `for_each_element` eliminates the
leading dimensions from consideration, and requires the user to
explicitly slice those as they see fit.

The downside is that the callback only gets a pointer and not a buffer,
so if a buffer object is required, the user needs to construct one
inside the callback.