Simplify some loop invariant code (#272)
This moves loop invariant stmts from the beginning of loops only out of
the loop. This does part of what is discussed in #270.

I decided to just add a TODO for more complicated forms of LICM, which
may not really be possible to have in the first place. I can't think of
a realistic way to have a loop invariant stmt consume a buffer that is
produced by a loop-variant stmt. Everything I can think of would do
weird things like read uninitialized memory.