Use ASSERT_THAT for simplify/substitute tests (#284)
Replace `test_simplify`, `test_substitute`, and similar helpers with
uses of ASSERT_THAT and a custom matcher.

The motivation for this: I'm working on a large change to the
simplifier, these tests are failing often and need a lot of debugging.
It's really annoying to figure out exactly which test is failing,
because the reported failure line is in `test_substitute` or similar. We
can't simply use ASSERT_EQ because we overload operator== for expr in a
different way. Calling `match` via an ASSERT_THAT matcher is easy and
solves all these problems.

There was similar discussion on #279 but I think this use is much
lighter weight.