Refactor buffer mutators to take an `src` buffer instead of always working in place (#286)
I discovered yet another challenging bug to fix with shadowing in the
simplifier, so I've decided to give up on supporting shadowing there.

In order to do this, we need to be able to rewrite programs to never
shadow, and to do that, we need to allow buffer mutators to take an src
buffer. This PR does that, but does not actually stop shadowing yet.
This PR reproduces existing behavior exactly (surprisingly), but with
the extra optional parameter to buffer mutators.

The next step will be to generate programs that do not have shadowing,
and remove the ability for the simplifier to understand shadowing, which
should be a big reduction in complexity and avoid probably many
undiscovered bugs. We will also want to rewrite programs to shadow
before evaluation, which is faster if buffer mutators work in place.