Add more structure into how crops are added (#287)
Currently, we have multiple places where crops are added and the
conditions on which they are added are pretty convoluted, which makes it
very hard to reason about how the IR will look like. In addition, the
current crop generation is not always correct, sometimes we generate too
many nested crops which can break storage folding or too few crops which
is reason for the test in this PR to fail at main.

In this PR, the crops are generated only inside of the crop_dim once per
loop, which should be sufficient, because the crops only need to be
recomputed after crop_dim. As a side benefit, the number of the added
crops should be O(N) instead of O(N^2). In addition, there were some
necessary changes in the storage folding to accomodate for the new crop
layout, most importantly the fold is now computed at the crop level and
not at the call site.

There are two diffs in the goldens: one seems to be pretty innocent and
the one is a slight regression where one of the inner loops is not
simplified away.