Enable optionally broadcasted copies to alias (#297)
This PR is a bunch of changes that enable optionally broadcasted copies
from #294 to alias when possible:
- `slice_buffer` and `slice_dim` can now slice out of bounds, which
makes the buffer base null. This makes it consistent with
`for_each_element` and related helpers, which is necessary to make
`copy_stmt` implementations correct when not doing any of the
optimizations for copies.
- Attempt to alias copies for the case that either the `src` or `dst` is
an allocation.
- Expand `is_copy`'s ability to understand these dynamic broadcast or
copy copies.
- A whole lot of refactoring of aliases/copy implementation passes. I
think there were a few bug fixes in here. I hope to clean this up
further still, this code is too hard to understand currently. I think it
can be improved a lot once I better understand what it's doing.

Now that we can do broadcasts properly and cheaply, I intend to go back
and clean out all the stride 0 hacks in the code in a follow-up change.