Clean up alias_buffers and remove broadcasting hacks (#298)
`alias_buffers` was the last thing I'm aware of that relied on
broadcasting hacks. This PR cleans up `alias_buffers` to hopefully make
it easier to understand (mainly by breaking the `copy_stmt` visitor into
separate cases for aliasing the src and dst, which are different), and
stops relying on the broadcasting hacks by correctly setting the bounds
of dimensions that we read broadcasts from. This PR then removes the
broadcasting hacks!