Change default strides to be determined by buffers/allocation callbacks instead of pipeline construction (#303)
Before this PR, default strides were set during pipeline construction.
This was flawed because if a user set some strides and not others, the
default generated strides would be confused. This issue is now fixed.

It additionally allows quite a few improvements:
- Allows the calling program more control over strides, such as to
control alignment/padding.
- Allows us to tell when strides are fixed vs. flexible. Right now,
aliased buffers will ignore the strides set by the user, potentially
invalidating them and violating assumptions the callbacks have made.
- The automatically generated strides algorithm is much more robust
(taken from
https://github.com/dsharlet/array/blob/38f8ce332fc4e26af08325ad0654c8516a445e8c/include/array/array.h#L897-L907,
but we don't need to accommodate compile time constants, so the
implementation is simpler).
- Very minor, but this removes some of the cruftiest expressions from
the programs we generate.