Improve sliding window and storage folding robustness (#320)
This started out by adding one little rule to canonicalize constants to
the RHS of less than, and it turned out that this unlocked a bunch of
improvements.

- Add rule to canonicalize constants to the RHS of <, which helps some
things simplify more.
- Add rules for finding bounds of `min(x, y) - max(z, w)`. When terms
from each side of the subtract are correlated, we might have a tighter
bound. This appears in fold factor expressions.
- Add `constant_upper_bound` helper, useful for fold factors.

The combination of these things means slide_and_fold_storage doesn't
need hacks to replace loop bounds with infinity any more.

This also fixes an issue identified by @vksnk: we currently allowed
folding to happen even when there was a consumer outside the loop we
fold in. We simply missed this by accident because we failed to slide
due to simplifier weakness.