Replace `truncate_rank` with `transpose` (#322)
I've been wanting to add a `transpose` buffer mutator for a while (to
make it more obvious when a `make_buffer` is a transpose in a large
pipeline), but didn't want to add a new one, because it makes writing
mutators a pain.

I realized that `truncate_rank` could be generalized a little bit to
become `transpose`, by replacing the number of dimensions with an array
of dimensions. `truncate_rank(x, rank, body)` is equivalent to
`transpose(x, {0, 1, ..., rank - 1}, body)`.

This could also use a placeholder to include any trailing dimensions,
but for now we don't have any simplifications that need that, because
`make_buffer` has no way to express such a thing.