Check that alias targets are big enough (#330)
This PR adds a check when considering a buffer for aliasing that the
aliasing target is big enough.

This required a bunch of new simplifier rules to enable proving that x
<= y where x, y are bounds of buffers.

Also fixes a bug where the replica pipeline generator dropped the
padding of copies.

This PR adds a few new tests that aren't passing the way they should yet
(the checks are commented with TODOs). It also disables the check for
aliasing in padded_stencil, which either needs major simplifier upgrades
to be able to prove is safe for aliasing, or needs the alias_buffers
pass to grow the allocation to the union of the original and the aliases
(I'm going to attempt that in my next aliasing PR).