Stronger simplification of nested selects (#334)
Something like this was originally added in #321, but reverted in #333
because it wasn't correct.

This PR adds something similar back, but it should be safe. This is
stronger than the rules that tried to do this, because there might be an
intervening op, and I don't want to add all the permutations of rules
that might be necessary for this.