Aliasing can now grow allocations when necessary (#331)
This is a major improvement to aliasing. Before this PR, aliasing of
padded allocations mostly just worked by accident. It relied on the fact
that padded allocations would be the outer allocation, and we'd alias
the smaller allocation to the padded allocation.

This PR fixes this to actually be safe and independent of the allocation
order (fixing #313). It now proves that buffers are big enough to
accommodate an alias, or if not, grows the allocation if possible.

This means we can reorder allocations to be more likely to be possible
to alias, which enables the broadcasted_elementwise test to alias.

This PR has some related changes and fixes:
- Added a new test that does a separable stencil with transposes in
between.
- Early frees handle shadowing
- Visualize implements shadowing properly
- Some cleanup of aliasing that hopefully makes it easier to understand.