Add rules for proving expressions are monotonic (#341)
I don't have super high confidence in these rules. I've fuzzed them a
ton, with intentionally broken rules to make sure there is coverage.
It's not the kind of thing I want to write a positive test for, because
coming up with the correct expression manually is difficult too. The way
I've been checking those is tedious.