Fuzz test each rule explicitly (#359)
This PR refactors the simplifier rules to make them available to a test
that generates expressions that match the rule exactly.

This testing revealed that we didn't handle boolean expressions well. We
would assume that for example, `x || true` is `x` , but this dropped an
implicit `x != 0 ? 1 : 0` in the original expression. So if `x` was
anything other than 0 or 1, the result of the expression changed. When
the expression was used by something that converted to bool, this didn't
matter. But we have rules that rewrite conditions into non-boolean
expressions, and that was a bug.

This PR fixes this by adding `x != 0` everywhere we convert `x` to a
boolean value, and the rewrite engine now asserts that if a pattern has
a boolean typed result, that the replacement is also a boolean typed
result.

Fixes #347