Remove shadowing in `simplify` (#369)
This PR removes shadowing support from the simplifier. Supporting
shadowing in the simplifier is very difficult, and there were still bugs
that weren't fixed. In fact, we relied on some of these bugs to
(incorrectly) simplify some of the code we simplified before this PR.

This works by adding a `deshadow` pass to rename shadowed symbols, and
an `optimize_symbols` that reshadows where possible (and in the future
should renumber symbols to be contiguous, #13 ).

Now `simplify` asserts that no symbol is shadowed, and removes shadowing
logic.

This reverts #305, which seems like the fix to #363, which I think only
works because it relies on simplifications that were buggy due to
shadowing.