Update loop bounds after sliding (#376)
This addresses the issue in #363.

The issue is that loop bounds may refer to buffer_min/buffer_max of some
buffer which are inferred by bounds inference and are outside of the
actual buffer bounds. This is a problem because the
buffer_min/buffer_max will be clamped to the actual buffer bounds during
evaluation (and maybe in some other places).