Fix LICM (#391)
I noticed that the `outer_product` tests all generated the exact same
body due to simplification of loops with only one statement in them. I
thought a quick and interesting thing to do would be to add another
stage to the pipeline, so the test covers more logic.

It turns out that `outer_product` is an interesting example of a
pipeline where you get loops with a mix of loop invariant and non-loop
invariant code, and the simplifier was broken in some of those cases.