Folded dimensions base points to 0 instead of min (#403)
This is a bit of a redo of #40 (reverted in #59), but instead of making
the base point to element 0 of every dimension, do it only for folded
buffers. Fixes #397

Because we only do this for folded buffers, and folded buffers now
always allocate `[0, extent)`, it avoids the problem of base pointers
being out of bounds or needing to be shifted from the allocation
pointer.

This resolves the long standing challenges with cropping folded buffers,
which doesn't work if the crop crosses a fold boundary and the offset is
non-zero (because that would require the folded offset to be negative
sometimes, which isn't possible because x % y >= 0).