Fix sanitizer ordering issues (#415)
The problem was that bounds of make_buffer's were using global variables
which are defined inside of the make_buffer body (incorrect because
bounds were using undefined variables). Later the simplifier would
substitute make_buffer bounds into lets which caused the incorrect
ordering of the lets (seems like it might be also a bug in simplifier).