Minor changes to support viewing traces in ui.perfetto.dev (#427)
It was failing to load a trace file because it didn't like the empty
event in the beginning. Also, for whatever reason it didn't like the
long thread ids, so had to add quotes for it to be able to handle
multi-threaded traces.

This is a matter of taste, but IMO ui.perfetto.dev has a better UI
compared to the built-in viewer and allows to get a lot of aggregated
statistics, etc, while still handling everything locally. The great
benefit though is that allows uploading and sharing traces with other
people which I think is extemely convinient. Example:
https://ui.perfetto.dev/#!/?s=17bcbcc16c0c8d6157bae9020e0c1d5e53a00db56841169fcdb9b885ea309e5e