Strengthen LICM (#428)
This PR makes LICM stronger, by allowing it to make non-loop-invariant
stmts loop invariant when it is safe to do so. It works by replacing the
loop variable with `expr()` only in crop bounds, and only if those crops
are not cropping a folded dimension. If there are other uses of the loop
variable (e.g. in a slice), the stmt will not be made to be loop
invariant.

This seems like a hack, but this kind of simplification has gotten us
really far already...