Fix deshadow slowness (#445)
`deshadow` was very slow due to two algorithmic complexity issues:
- The way `substitute` handled shadowing was very inefficient: tracking
all shadowed variables, and checking at substitute time if a
substitution should be disallowed due to shadowing. Now, we stop
traversing the IR as soon as any possible substitutions become shadowed,
which also avoids the expensive way we tracked shadowed variables
(linear search in a vector).
- `insert_unique` used a linear search to find an unused symbol name,
and a linear search to check if a symbol name was used, both are fixed.

The test added in this PR took ~3 seconds before these fixes, and now
takes 150ms, where almost all the time is spent in `insert_unique`. It
could be improved further, but not easily, and it's not worth doing for
now.