Fix incorrect aliasing for translated copies (#456)
When we tried to expand allocations to accommodate a possibly padded
alias, we used the "native" coordinates for both buffers, which is
incorrect if there is a translation between the aliases. This PR fixes
that.

Some related notes:
- The old "replace copy with pad" logic was pretty broken, and also
pointless, because pad simply called copy with an aliased buffer
anyways, expecting the copy parts of the operation to be a no-op. We
might want to optimize this at some point, but it's low priority for the
moment.
- The asserts added in #455 were overzealous if the buffer is empty,
fixed this to allow empty buffers.