Make `expr(const base_expr_node*)` and `stmt(const base_stmt_node*)` explicit (#482)
This is a messy change. The core of it is as the title says: make
`expr(const base_expr_node*)` and similar for `stmt` `explicit`. This
uncovered a few places where we were doing silly things, like
constructing an `expr` just to call a helper function that doesn't
affect the lifetime of its argument.

The rest is a wrapper to attempt to avoid requiring explicit conversions
everywhere, but without the excessive reference counting due to
constructing temporary `expr` objects.

This speeds up the simplifier by 5-10%, and *should* reduce code size,
but I think the compiler is inlining more than before this change, so
it's a wash.