Ensure that the plan is initialized for zero-D buffers in make_for_each_loops_impl() (#496)
This is really a sanitizer fix: if the buffers are zero-D then we left
the `plan` filled with stack garbage. This usually didn't bite us
because we would just add random garbage to a pointer that was never
used afterwards, but (e.g.) UBSAN has better checking for
pointer-arithmetic-overflow, so ensuring that our strides are 0 prevents
sanitizer failures.