Upgrade as_constant and as_variable (#500)
They now return `std::optional<T>` instead of `const T*`. This turns out
to matter in subtle cases like

```
      int foo = *as_constant(simplify(thing()));
```

where the return value of `simplify` is a temporary that's discarded
immediately, but that as_constant() assumes will remain valid. Using
`std::optional` captures the value in a way that should be safer against
this pattern, as well as more easily debuggable (since dereferencing an
empty std::optional is caught by clang-tidy in many cases and is also
checked for by debug builds of libc++).