compute_allocation_bounds() should handle all input_syms_
In theory, well-formed input to Slinky should only provide `inputs` that are actually inputs to at least one `func` in the pipeline; in practice, translation layers can get this wrong (e.g. constant tensors that are marked as external inputs), and if we end up with anything in `input_syms_` that doesn't have `allocation_bounds_` filled in, things go poorly. Fortunately, we can reasonably assume that everything in `input_syms` that fits this case has well-known boundaries that we can simply copy, making it simple to be a bit permissive in this case.