Add make_binary_maybe_constant() (#505)
In simplify(), we need to be more careful about operations that can
overflow; in particular, add/sub call make_binary() which can overflow
in degenerate conditions. This PR adds make_binary_maybe_constant(),
which detects overflow for constants and in those cases preserves the
expression in `expr` form to avoid overflow.