Replace buffer intrinsics with `variable` fields (#530)
This PR changes buffer intrinsics (e.g. `buffer_min(buf, dim)`) from
being implemented as `call` nodes to `variable` nodes, by adding new
fields to `variable`.

This change has a few motivations and implications:
- It is an improvement to #396 and #466, because evaluating a buffer
intrinsic now requires evaluating only one node instead of 3. The
`BM_buffer_meta` benchmark gets ~5x faster with this branch.
- It loses flexibility, because the `buf` and `dim` arguments to buffer
intrinsics are now required to be constants. However, in dozens of
places in the code, we asserted those values were `variable` and
`constant` nodes anyways, and I doubt we'd ever change that.
- The main motivation for this change is that the simplifier currently
treats variables and buffers very differently, with different features
for each. `expr_info` tracks bounds and alignment for each variable, but
`buffer_info` only tracks the value of the various buffer fields. This
PR doesn't do this yet, but I have a follow-up WIP branch that (mostly)
unifies this. This has a bunch of benefits:
- It's a step towards passing buffer information to the simplifier from
outside, currently we cannot. We could add new arguments for this, but
it's a plumbing mess to do so.
- It allows us to learn bounds of buffer fields, currently we can only
learn specific values. Right now, if the program does
`check(buffer_min(x, 0) == 0 && buffer_max(x, 0) > 0)`, we can learn
that the buffer min is 0 (because it is a specific value we can store in
`buffer_info`), but we cannot learn that the `buffer_max` (and thus the
buffer extent too) is greater than 0, which would be valuable
information in many cases.
  - The simplifier code gets simpler.