Small optimizations for very lightweight pipelines (#544)
This remove special case from `init_strides` that only exists to make
strides "look" better. Some things assert strides are the product of a
previous dimension extent and stride, even if the extent is 1 (so the
stride is irrelevant to anything). If people want to call other programs
with such asserts, they can fix up the strides themselves, but we
shouldn't pay the cost to fix them up when they don't matter.

Also:
- Avoid calling `free` on `nullptr` (in case it isn't inline)
- Inline some things in `evaluate` that proved helpful.