Only alias copies, not calls (#548)
This changes `alias_buffers` to only alias copies, rather than copies
and in-place calls. Doing in-place calls correctly is tricky, and
doesn't really need all the complexity of `alias_buffers`. I think it
will be better done as a separate pass. We also don't have much that
actually relies on this functionality, and in fact a few tests were
accidentally relying on it (fixed in this PR), and this should unblock
#543.