Change IR generation to minimize the depth of the allocation nesting (#543)
This attempts to lay out allocation nodes such that the nesting is
minimized. The general idea is to iteratively build up a tree of
allocations starting from the allocations with the allocations with the
shortest life time as the lowest level of the tree. This is not always
possible in general to do and there are corner cases where nesting of
allocations will have depth N regardless of the approach, but in most
practical situations this will produce a structure close to the tree
(for example, for the linear pipeline it should build a perfect tree of
depth ~log(N)).

There are a couple of tests in the copy_pipeline which don't alias
anymore, so the checks there are commented out until simplifier is able
to use buffer bounds better.

---------

Co-authored-by: Dillon <dsharlet@gmail.com>