Fix `canonicalize_nodes` for similar-looking calls (#566)
We had a terrible bug where if we found two calls with the same
arguments (which can happen due to aliasing and reusing allocations), we
thought they were equivalent, and canonicalized them.

However, `match` did not check if `call_stmt::target` was equal, which
caused us to possibly call a different function.

This also fixes `fuse_siblings` to recursively mutate. Fusing a sibling
can reveal new fusions in the result that we missed.