Add `constant_buffer` node (#569)
Before this PR, we didn't really have a way to handle constant buffers
properly. We'd search for them in user programs, and pass them
essentially as any other "input" buffer to the pipeline. This is a
little bit of a headache.

This PR introduces the `constant_buffer` node. It's similar to
`make_buffer`, but it owns a (shared) pointer to a `raw_buffer`, so it
can extend the lifetime of an actual buffer. It also can be evaluated
more efficiently.