Optimize `for_each_element` (#595)
I've now invested a silly amount of effort in this function, and I'm
hoping this is basically as good as it can get, barring the possibility
that we change the definition of `raw_buffer` in some way that affects
this. This makes it 10-30% faster, depending on the benchmark.

The major change here is to use function pointers stored in the
`for_each_loop` stack, instead of flags + a dispatcher function. Also:
- Add the `constexpr` `Extent` template parameter of our `std::span`
clone, and use it to avoid passing an unused parameter around.
- Avoid changing the type of rank/dim indices, which adds some overhead.
- A bunch of other minor simplifications and cleanups.

My ultimate goal was to get to the point where we could use computed
gotos and implement this without function calls at all. However, I
wasn't actually able to beat the function pointer approach, because my
manual stack management is basically equivalent to what the standard
calling convention does, if not a little worse, and it's a lot more of a
headache to debug it.

This change also changes the benchmarks to be no-ops, instead of
`memcpy`/`memset` calls. I found that there were some optimizations that
made `for_each_element` faster, but were overall slower because of
compiler stupidity in the callbacks that I don't think are durable
across real usages (dropping from 4 register passed parameters to 3
should never hurt).