Fix the order of the function during the IR construction. (#608)
Apparently, the test which checked compute_at was disabled and was
failing at head after I enabled it. The issue is that we first were
visiting the inner loop which might violate the assumption that the
function are added in the order of their occurance in the final IR (this
is important so we can merge lifetimes correctly).