Fix store_at and add a test for it (#610)
This PR the part which produces allocate nodes from ```build``` to
```make_loop```, which makes it more logical because allocations are
tied to the loop level and make_loop has view of all statements included
in it. In order to make code simpler, we now start IR construction by
creating a "root" loop in which case we only produce the body of the
"loop", but not the crops or actual loop.

(Technically, this got broken by #608, but since that was the fix
```compute_at``` which wasn't working, it's fair to say that
```store_at``` didn't work either by transitivity)