Add `slice` constant for use with `at`/`address_at` (#611)
This allows calling `at` and `address_at` with a value `slice` that
essentially skips the dimension. This is useful for writing func
implementations where you want to get a pointer to the beginning of a
region to process that is the min in some dimensions (that aren't the
training dimensions) and not others.

It's a bit tricky what this means for folded dimensions, because the
base doesn't point to the min of folded dimensions. I think the right
thing to do is just skip these dimensions, because the code is likely to
add the offset for this dimension later.