A more targeted store_at test and fix for it (#626)
Previously, we would produce crop_buffer-s for all buffers even if they
are allocated inside of the body of the crop. In this case we really
need only buffer metadate, so make_buffer is enough.

This also regresses back the previosly improved test. I think we need to
teach slide_and_fold_storage to take bounds from make_buffer as well to
fix it -- will fix in follow up CL.