Rewrite copy, pad, fill to always take buffers instead of scalar values (#633)
- Remove `fill`, because it is equivalent to `copy` with a scalar buffer
as the `src`.
- Change the padding argument of `copy` and `pad` to be a buffer instead
of a scalar value. This allows the padding to vary across the output.

This PR is a fairly significant performance regression for some cases of
`fill` and `copy` (`fill`/`copy` is 1.2-2x slower for broadcasting
copies, only slightly slower for dense copies). It is also an
improvement in some cases (copying from folded dense dimensions).

However, we need the extra flexibility enabled by this, and I believe we
could recover the performance loss, but it would require handling quite
a few cases ({padding, no padding} x {src broadcast, src dense, src
strided} x {pad broadcast, pad dense, pad strided}). I think this is
reasonably doable, but a bit messy. We can wait until we see the
performance of the current approach is an issue before addressing this.

This PR is part 1 of at least 2. The next step will be to change the
`padding` argument of `func` to also be a buffer. That argument is
unchanged by this PR (the implementation wraps that scalar argument with
a buffer to use the new `copy`/`pad` implementation).